need to do 'p' flag

+++ libft/nbr_size.c
@@ -6,13 +6,13 @@
 /*   By: narajaon <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/07/24 12:13:57 by narajaon          #+#    #+#             */
-/*   Updated: 2017/07/24 12:16:03 by narajaon         ###   ########.fr       */
+/*   Updated: 2017/07/25 15:04:56 by narajaon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		nbr_size(unsigned int nbr)
+int		nbr_size(unsigned long long int nbr)
 {
 	if (nbr < 10)
 		return (1);
